requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.9.1
pytz==2021.1
requests==2.25.1
simplejson==3.17.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
